requirements.txt
amqp==1.4.9
ansicolors==1.0.2
anyjson==0.3.3
appdirs==1.4.3
billiard==3.3.0.23
bleach==2.0.0
celery==3.1.24
cffi==1.9.1
chardet==2.3.0
colorama==0.3.2
configparser==3.5.0b2
coverage==4.0.3
cryptography==1.7.1
defusedxml==0.4.1
Django==1.8.17
django-admin-tools==0.7.1
django-braces==1.9.0
django-celery==3.1.17
django-colorful==1.2
django-contrib-comments==1.7.3
django-cryptographic-fields==0.3.0
django-easy-audit==0.8
django-extensions==1.5.9
django-filer==1.0.6
django-filter==0.15.3
django-flat-theme==1.1.3
django-haystack==2.4.1
django-hstore==1.4.2
django-leaflet==0.22.0
django-model-utils==3.0.0
django-mptt==0.8.6
django-phonenumber-field==0.7.2
django-polymorphic==0.8.1
django-softdelete==0.7.0
django-taggit==0.22.0
django-taggit-autosuggest==0.3.0
django-threadedcomments==1.0.1
djangorestframework==3.3.2
-e git+git://github.com/djangonauts/django-rest-framework-gis.git@master#egg=djangorestframework-gis
djangorestframework-xml==1.3.0
easy-thumbnails==2.3
env-tools==2.0.0
futures==3.0.3
html5lib==0.999999999
httpie==0.9.3
httplib2==0.9.2
idna==2.1
jsonpickle==0.9.3
kombu==3.0.37
model-mommy==1.3.2
oauth2client==4.0.0
packaging==16.8
phonenumbers==8.6.0
Pillow==3.0.0
psutil==3.3.0
psycopg2==2.6.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.17
Pygments==2.1
pyparsing==2.1.10
pysolr==3.4.0
python-magic==0.4.10
pytz==2016.7
requests==2.9.1
rsa==3.4.2
six==1.10.0
tini==3.0.1
Unidecode==0.4.18
urllib3==1.9.1
virtualenv==1.11.6
